Two scripts made
FireBallStats extends ProjectileWeapon, FireBall contains monobehaviour for fireball gameobject
