+++ Assets/Other/FirePoint.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e3d80a07660953c4a88f18c487e93e52
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
