+++ Assets/Scripts/FireBall.cs.meta
@@ -3,7 +3,9 @@ guid: 10d5361225b4bca44823f23e0ed28c57
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
-  defaultReferences: []
+  defaultReferences:
+  - fireball: {fileID: 11400000, guid: 7e788612f142c254ab569a4d16014179, type: 2}
+  - fire: {fileID: 9000563240690001610, guid: 797597cd0f11fe64dac26bcc59b2d76c, type: 3}
   executionOrder: 0
   icon: {instanceID: 0}
   userData:
