+++ Assets/Scripts/FireBallStats.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FireBallStats : ProjectileWeapon
+{
+    public override void Fire()
+    {
+
+    }
+
+    public override void Charge()
+    {
+
+    }
+
+}
