+++ Assets/Scripts/FireBallStats.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 85788614b0dad3b43a1c38f01361fd1e
+guid: 3d0236914225847469bc1c05ceda762a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
