+++ Assets/Scripts/Prefabs/Data.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3cc4fbe6fee96224fa50447875b3b210
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
