+++ Assets/Scripts/Prefabs/FireBall.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 9000563240690001615}
   - component: {fileID: 9000563240690001610}
   - component: {fileID: 9000563240690001613}
+  - component: {fileID: 183451014055268008}
   m_Layer: 0
   m_Name: FireBall
   m_TagString: Untagged
@@ -49,7 +50,7 @@ SpriteRenderer:
   m_ReflectionProbeUsage: 1
   m_RayTracingMode: 0
   m_RayTraceProcedural: 0
-  m_RenderingLayerMask: 1
+  m_RenderingLayerMask: 4294967295
   m_RendererPriority: 0
   m_Materials:
   - {fileID: 2100000, guid: a97c105638bdf8b4a8650670310a4cd3, type: 2}
@@ -72,7 +73,7 @@ SpriteRenderer:
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: 0
   m_SortingLayer: 0
-  m_SortingOrder: 0
+  m_SortingOrder: 3
   m_Sprite: {fileID: 10913, guid: 0000000000000000f000000000000000, type: 0}
   m_Color: {r: 0.9716981, g: 0.07588786, b: 0, a: 1}
   m_FlipX: 0
@@ -121,3 +122,16 @@ CircleCollider2D:
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
   m_Radius: 0.1
+--- !u!114 &183451014055268008
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9000563240690001614}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 10d5361225b4bca44823f23e0ed28c57, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  fireballStats: {fileID: 11400000, guid: 3cc4fbe6fee96224fa50447875b3b210, type: 2}
