--- Assets/Scripts/Shoot.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Shoot : MonoBehaviour
-{
-    public Rigidbody2D rb;
-    public float speed = 20f;
-    ProjectileWeapon fireball;
-
-    void Start()
-    {
-        rb.velocity = transform.right * speed;
-    }
-
-    void OnTriggerEnter2D(Collider2D hitInfo)
-    {
-        
-        Destroy(gameObject);
-    }
-}
